requirements.txt
anaconda==0.0.1.1
annotated-types==0.6.0
anyio==4.3.0
blinker==1.8.2
certifi==2024.2.2
click==8.1.7
colorama==0.4.6
deprecation==2.1.0
exceptiongroup==1.2.1
Flask==3.0.3
Flask-Cors==4.0.1
geojson==3.1.0
gotrue==2.4.2
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.0
postgrest==0.16.4
psycopg2==2.9.9
pydantic==2.7.1
pydantic_core==2.18.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
realtime==1.0.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.30
storage3==0.7.4
StrEnum==0.4.15
supabase==2.4.5
supafunc==0.4.5
typing_extensions==4.11.0
waitress==3.0.0
websockets==12.0
Werkzeug==3.0.3
anaconda==0.0.1.1
annotated-types==0.6.0
anyio==4.3.0
blinker==1.8.2
certifi==2024.2.2
click==8.1.7
colorama==0.4.6
deprecation==2.1.0
exceptiongroup==1.2.1
Flask==3.0.3
Flask-Cors==4.0.1
geojson==3.1.0
gotrue==2.4.2
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.0
postgrest==0.16.4
psycopg2==2.9.9
pydantic==2.7.1
pydantic_core==2.18.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
realtime==1.0.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.30
storage3==0.7.4
StrEnum==0.4.15
supabase==2.4.5
supafunc==0.4.5
typing_extensions==4.11.0
waitress==3.0.0
websockets==12.0
Werkzeug==3.0.3
